+++ omc3/model/accelerators/accelerator.py
@@ -6,6 +6,7 @@ This module provides high-level classes to define most functionality of ``model.
 It contains entrypoint the parent `Accelerator` class as well as other support classes.
 """
 import re
+import os
 from pathlib import Path
 from typing import List, Union, Sequence
 
@@ -16,7 +17,7 @@ from generic_parser.entrypoint_parser import EntryPointParameters
 
 from omc3.model.constants import (
     ERROR_DEFFS_TXT,
-    JOB_MODEL_MADX,
+    JOB_MODEL_MADX_NOMINAL,
     MODIFIER_TAG,
     MODIFIERS_MADX,
     TWISS_AC_DAT,
@@ -59,6 +60,8 @@ class Accelerator:
         AccElementTypes.ARC_BPMS: r".*",
     }
     BPM_INITIAL = "B"
+    NAME=None
+    REPOSITORY=None
 
     @staticmethod
     def get_parameters():
@@ -95,11 +98,11 @@ class Accelerator:
         params.add_parameter(
             name="energy",
             type=float,
-            help="Energy in Tev.",
+            help="Energy in GeV.",
         )
         params.add_parameter(
             name="modifiers",
-            type=Path,
+            type=PathOrStr,
             nargs="*",
             help="Path to the optics file to use (modifiers file).",
         )
@@ -118,6 +121,7 @@ class Accelerator:
         self.model_best_knowledge = None
         self.elements = None
         self.error_defs_file = None
+        self.acc_model_path = None
         self.modifiers = None
         self._beam_direction = 1
         self._beam = None
@@ -137,11 +141,7 @@ class Accelerator:
         else:
             self.init_from_options(opt)
 
-    def init_from_options(self, opt) -> None:
-        if opt.nat_tunes is None:
-            raise AcceleratorDefinitionError("Argument 'nat_tunes' is required.")
-        if (opt.drv_tunes is None) and (opt.driven_excitation is not None):
-            raise AcceleratorDefinitionError("Argument 'drv_tunes' is required.")
+    def init_from_options(self, opt):
         self.nat_tunes = opt.nat_tunes
 
         if opt.driven_excitation is not None:
@@ -192,6 +192,17 @@ class Accelerator:
         best_knowledge_path = model_dir / TWISS_BEST_KNOWLEDGE_DAT
         if best_knowledge_path.is_file():
             self.model_best_knowledge = tfs.read(best_knowledge_path, index="NAME")
+
+        # Base Model ########################################
+        if self.REPOSITORY is not None:
+            acc_models = model_dir / self.REPOSITORY
+
+            if acc_models.is_dir():
+                if acc_models.is_symlink():
+                    self.acc_model_path = Path(os.readlink(acc_models)).absolute()
+                else:
+                    self.acc_model_path = acc_models
+            # else this wasn't an acc-models based model
 
         # Modifiers #########################################
         self.modifiers = _get_modifiers_from_modeldir(model_dir)
@@ -256,7 +267,12 @@ class Accelerator:
         """
         Verifies that this instance of an `Accelerator` is properly instantiated.
         """
-        raise NotImplementedError("A function should have been overwritten, check stack trace.")
+        # since we removed `required` args, we check here if everything has been passed
+        if self.model_dir is None:
+            if self.nat_tunes is None:
+                raise AttributeError("Natural tunes not set (missing `--nat_tunes` flag?)")
+            if self.excitation != AccExcitationMode.FREE and self.drv_tunes is None:
+                raise AttributeError("Driven excitation selected but no driven tunes given (missing `--drv_tunes` flag?)")
 
     def get_exciter_bpm(self, plane: str, commonbpms: List[str]):
         """
@@ -355,13 +371,16 @@ class AcceleratorDefinitionError(Exception):
 
 def _get_modifiers_from_modeldir(model_dir: Path) -> List[Path]:
     """Parse modifiers from job.create_model.madx or use modifiers.madx file."""
-    job_file = model_dir / JOB_MODEL_MADX
+    job_file = model_dir / JOB_MODEL_MADX_NOMINAL
     if job_file.exists():
         job_madx = job_file.read_text()
 
         # find modifier tag in lines and return called file in these lines
+        # the modifier tag is used by the model creator to mark which line defines modifiers
+        # see e.g. `get_base_madx_script()` in `lhc.py`
+        # example for a match to the regex: `call, file = 'modifiers.madx'; MODIFIER_TAG`
         modifiers = re.findall(
-            fr"\s+call,\s*file\s*=\s*[\"\']?([^;\'\"]+)[\"\']?\s*;\s*{MODIFIER_TAG}",
+            fr"\s*call,\s*file\s*=\s*[\"\']?([^;\'\"]+)[\"\']?\s*;\s*{MODIFIER_TAG}",
             job_madx,
             flags=re.IGNORECASE,
         )